Refactor and fix Rewind policy
Previously, if one of the messages that deactivated a policy was received before
the policy was activated, it would still activate the policy even though
it shouldn't.

Also, multiple received() statements weren't actually composed, merely
stacked.

This commit refactors the policy code into three phases: declaration,
build, and installation.
This way, the received() statement merely constructs a list of
expectations in disjunction normal form, which are then turned into
reactors during the build phase, and added to the adapter after that.